move DatabaseMiddleware type out of KnexMiddleware
