+++ packages/backend/src/peripherals/database/DatabaseMiddleware.ts
@@ -1,0 +1,7 @@
+import { Knex } from 'knex'
+
+export interface DatabaseMiddleware {
+  add: (cb: () => Promise<void>) => void
+  execute: () => Promise<void>
+}
+export type DatabaseTransaction = Knex.Transaction
